+++ apps/wallet-mobile/src/features/Receive/useCases/DescribeSelectedAddressScreen.tsx
@@ -17,7 +17,10 @@ import {useMultipleAddressesInfo} from '../common/useMultipleAddressesInfo'
 import {useNavigateTo} from '../common/useNavigateTo'
 import {useReceiveAddressesStatus} from '../common/useReceiveAddressesStatus'
 import {useStrings} from '../common/useStrings'
-import {QRs} from '../illustrations/QRs'
+import {
+  SingleOrMultipleAddressesModal,
+  singleOrMultipleAddressesModalHeight,
+} from '../common/SingleOrMultipleAddressesModal/SingleOrMultipleAddressesModal'
 
 export const DescribeSelectedAddressScreen = () => {
   const strings = useStrings()
@@ -44,7 +47,7 @@ export const DescribeSelectedAddressScreen = () => {
     isSingle &&
       isMultipleAddressesUsed &&
       isShowingMultipleAddressInfo &&
-      openModal(strings.singleOrMultiple, <Modal />, modalHeight)
+      openModal(strings.singleOrMultiple, <SingleOrMultipleAddressesModal />, singleOrMultipleAddressesModalHeight)
   }, [isShowingMultipleAddressInfo, isSingle, isMultipleAddressesUsed, openModal, strings.singleOrMultiple])
 
   useFocusEffect(
@@ -89,54 +92,6 @@ export const DescribeSelectedAddressScreen = () => {
   )
 }
 
-const modalHeight = 580
-const Modal = () => {
-  const {styles, colors} = useStyles()
-  const strings = useStrings()
-  const {enableMultipleMode} = useAddressModeManager()
-
-  const {hideMultipleAddressesInfo} = useMultipleAddressesInfo()
-
-  const {closeModal} = useModal()
-
-  const handleOnMultiple = () => {
-    enableMultipleMode()
-    handleOnSingle()
-  }
-
-  const handleOnSingle = () => {
-    hideMultipleAddressesInfo()
-    closeModal()
-  }
-
-  return (
-    <View style={styles.modal}>
-      <QRs />
-
-      <Text style={[styles.details, {color: colors.details}]}>{strings.singleOrMultipleDetails}</Text>
-
-      <Spacer fill height={24} />
-
-      <View style={styles.buttonContainer}>
-        <Button
-          outline
-          title={strings.selectMultiple}
-          textStyles={{
-            color: colors.details,
-          }}
-          onPress={handleOnMultiple}
-        />
-
-        <Spacer height={6} />
-
-        <Button shelleyTheme title={strings.singleAddressWallet} onPress={handleOnSingle} style={styles.button} />
-      </View>
-
-      <Spacer height={24} />
-    </View>
-  )
-}
-
 const useStyles = () => {
   const {theme} = useTheme()
 
@@ -145,19 +100,6 @@ const useStyles = () => {
       flex: 1,
       backgroundColor: theme.color.gray.min,
       padding: 16,
-    },
-    modal: {
-      flex: 1,
-      backgroundColor: theme.color['bottom-sheet-background'],
-      alignItems: 'center',
-      justifyContent: 'space-between',
-    },
-    buttonContainer: {
-      alignSelf: 'stretch',
-      backgroundColor: theme.color.gray.min,
-    },
-    details: {
-      ...theme.typography['body-1-l-regular'],
     },
     address: {
       flex: 1,
@@ -170,7 +112,6 @@ const useStyles = () => {
 
   const colors = {
     requestSpecificAmountTextColor: theme.color.primary[500],
-    details: theme.color.gray[900],
   }
 
   return {styles, colors} as const
